proyectos/telar-almas/requirements.txt
# El Telar de las Almas - Requerimientos
supabase
flatlib
pydantic
python-dotenv
requests
